move configuration logic to separate file
