+++ pkg/autoscaling/engines/containership/containership.go
@@ -1,7 +1,6 @@
 package containership
 
 import (
-	"encoding/json"
 	"math/rand"
 	"os"
 	"time"
@@ -31,13 +30,6 @@ type Engine struct {
 	nodeLister corelistersv1.NodeLister
 	cloud      cscloud.Interface
 	config     *cloudConfig
-}
-
-type cloudConfig struct {
-	Address         string
-	TokenEnvVarName string
-	OrganizationID  string
-	ClusterID       string
 }
 
 // NewClient creates a new instance of the containership AutoScaling Engine, or an error
@@ -120,65 +112,4 @@ func (e Engine) scaleStrategyRandom(nodePoolID string, numNodes int) (bool, erro
 	}
 
 	return true, nil
-}
-
-func (c *cloudConfig) defaultAndValidate(configuration map[string]string) error {
-	// Round trip the config through JSON parser to populate our struct
-	j, _ := json.Marshal(configuration)
-	json.Unmarshal(j, c)
-
-	if err := c.defaultAndValidateAddress(); err != nil {
-		return err
-	}
-
-	if err := c.defaultAndValidateTokenEnvVarName(); err != nil {
-		return err
-	}
-
-	if err := c.defaultAndValidateOrganizationID(); err != nil {
-		return err
-	}
-
-	if err := c.defaultAndValidateClusterID(); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (c *cloudConfig) defaultAndValidateAddress() error {
-	if c.Address == "" {
-		c.Address = "https://provision.containership.io"
-	}
-
-	return nil
-}
-
-func (c *cloudConfig) defaultAndValidateTokenEnvVarName() error {
-	if c.TokenEnvVarName == "" {
-		return errors.New("tokenEnvVarName must be provided")
-	}
-
-	token := os.Getenv(c.TokenEnvVarName)
-	if token == "" {
-		return errors.New("unable to get Containership Cloud API cluster token")
-	}
-
-	return nil
-}
-
-func (c *cloudConfig) defaultAndValidateOrganizationID() error {
-	if c.OrganizationID == "" {
-		return errors.New("organizationID must be provided")
-	}
-
-	return nil
-}
-
-func (c *cloudConfig) defaultAndValidateClusterID() error {
-	if c.ClusterID == "" {
-		return errors.New("clusterID must be provided")
-	}
-
-	return nil
 }
